refactor: 폴더 관심사별로 재분류 및 그에 따른 경로 변경
pages 폴더에는 페이지 파일만 있도록 이동
components 폴더 하위에 Restaurant, RestaurantDetail, Restaurants 폴더 추가
