+++ src/components/Categories/CategoriesContainer.jsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import {
   selectCategory,
   loadRestaurants,
-} from '../../slice/slice';
+} from '../../slice/restaurantSlice';
 
 import { get } from '../../utils';
