+++ src/components/Restaurant/RestaurantContainer.test.jsx
@@ -82,7 +82,7 @@ describe('RestaurantContainer', () => {
           fireEvent.change(getByLabelText(label), { target: { value } });
 
           expect(dispatch).toBeCalledWith({
-            type: 'application/changeReviewField',
+            type: 'reviewSlice/changeReviewField',
             payload: { name, value },
           });
         });
