+++ src/pages/RestaurantPage/RestaurantPage.jsx
@@ -1,6 +1,6 @@
 import { useParams } from 'react-router-dom';
 
-import RestaurantContainer from './RestaurantContainer';
+import RestaurantContainer from '../../components/Restaurant/RestaurantContainer';
 
 export default function RestaurantPage({ params }) {
   const { id } = params || useParams();
